Split core/data.py into individual files in core/data module
